requirements.txt
streamlit>=1.28.0
openai>=1.3.0
pydantic>=2.5.0
reportlab>=4.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pypdf>=3.17.0


